+++ resources/js/components/organisms/LoginForm.tsx
@@ -1,39 +1,16 @@
-import { useState } from "react";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
-import { useNavigate } from "react-router-dom";
-import { ApiClient } from "../../apis/ApiClient";
+import { useLogin } from "../../hooks/useLogin";
 
 export const LoginForm = () => {
-    const [inputEmail, setInputEmail] = useState("");
-    const [inputPassword, setInputPassword] = useState("");
-    const [error, setError] = useState(false);
-    const navigate = useNavigate();
-
-    const handleLogin = async (e) => {
-        e.preventDefault();
-        try {
-            const { data } = await ApiClient.get("/sanctum/csrf-cookie")
-                .then(() => {
-                    return ApiClient.post("/api/user/login", {
-                        email: inputEmail,
-                        password: inputPassword,
-                        // email: "test@example.com",
-                        // password: "password",
-                    });
-                })
-                .catch((err) => {
-                    setError(true);
-                    throw err;
-                });
-            if (data.isAuth) {
-                //グローバルにログイン状態をセット
-                navigate("/");
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    };
+    const {
+        inputEmail,
+        inputPassword,
+        error,
+        setInputEmail,
+        setInputPassword,
+        handleLogin,
+    } = useLogin();
     return (
         <Form onSubmit={(e) => handleLogin(e)}>
             {error && (
